+++ buildSrc/src/main/java/org/opensearch/gradle/testclusters/TestClusterConfiguration.java
@@ -55,6 +55,8 @@ public interface TestClusterConfiguration {
 
     void setTestDistribution(TestDistribution distribution);
 
+    void extension(ExtensionsProperties extension);
+
     void plugin(Provider<RegularFile> plugin);
 
     void plugin(String pluginProjectPath);